requirements.txt
discord.py
praw
prawcore
markovify
hackerrank-sdk
